Try to look up source/dest feature stores for proper naming.
When generating SQL joins, app-schema assumes the source *database
schema* is the same as the destination *database schema* (same
JDBCDataStore). This isn't necessarily the case and it's perfectly
legitimate to join tables in different schemas. This patch attempts to
look up the appropriate delegate stores before attempting to encode
their names.